Add a map to DynamicLoader to bypass dlsym
We *may* be seeing issues related to dlopen/dlsym on a Vortex86DX3.
This will allow bypassing dlopen/dlsym when looking up symbols, in
order to test that theory.

The NIDAS_CREATOR_FUNCTION macros have been expanded to define
a register function for that symbol, which is declared
__attribute__((constructor)) so that it is executed on library load.
The register function adds the create function to the DynamicLoader map.