still tweaking timing on serial_sensor test
sleep in sensor_sim after finishing data and before unlinking the pty, in case
that helps dsm read the last sample from the pty before it's gone.  also
increase the sleep after dsm quits, in case that alleviates failures where the
dsm_server does not receive the last sample from the dsm.  this is all because
of test failures on my fedora laptop which had not failed before...