allow SampleTag Site to change
A SampleTag should get it's Site from the containing Sensor if available,
rather than always from the DSMConfig.  Also, whenever the DSMSensor changes,
propagate that to all the derived properties that are derived from it.
Consolidate the two constructors into one with a default argument.  Add
SampleTag::toString() as a logging convenience.