allow site attribute in a sensor
DSMSensor now accepts a site attribute to assign that sensor to a different
site than the dsm, which mostly means that SampleTag and Variable properties
can be derived from a different site than the dsm.  Use the new DOMable api
changes to collect attributes.  Add tests for the DOMable API and the handling
of the site attribute.  There is still more to do migrating fromDOMElement()
implementations to the extended DOMable api.